WJ-102 - Update: Move EntityImage to containers;
