+++ src/components/index.ts
@@ -3,7 +3,6 @@ export { ColumnLayout, RowLayout, Wrapper, Container } from 'components/Layout';
 export { default as Button } from './Button';
 export { default as EnvironmentLabel } from './EnvironmentLabel';
 export { default as Input } from './Input';
-export { default as EntityImage } from './EntityImage';
 export { default as MovieHighlight } from './MovieHighlight';
 export { default as Person } from './Person';
 export { default as Profile } from './Profile';
