+++ src/containers/EntityImage/dtos/ContainerProps.ts
