+++ src/containers/EntityImage/dtos/index.ts
