+++ src/containers/EntityImage/styles.ts
