+++ src/containers/MovieList/dtos/index.ts
@@ -1,6 +1,6 @@
 import DefaultProps from 'shared/dtos';
-import MovieProps from 'components/EntityImage/dtos';
-import FavoriteProps from 'domains/Favorites/api/List/Response';
+// import MovieProps from 'containers/EntityImage/dtos';
+// import FavoriteProps from 'domains/Favorites/api/List/Response';
 import { Color } from 'shared/enums';
 
 export default interface Props extends DefaultProps {
