+++ src/containers/index.ts
@@ -1,3 +1,4 @@
+export { default as EntityImage } from './EntityImage';
 export { default as Filmography } from './Filmography';
 export { default as Footer } from './Footer';
 export { default as Header } from './Header';
